We like bf16

+++ MaxText/layers/pipeline.py
@@ -141,7 +141,7 @@ class Pipeline(nn.Module):
     def replace_with_zeros(pytree):
       def _replace_with_zero_array(leaf):
         all_repeats = jnp.zeros_like(leaf)
-        #all_repeats = all_repeats.astype(jnp.bfloat16) # TODO: rawr
+        all_repeats = all_repeats.astype(jnp.bfloat16) # TODO: rawr
         return all_repeats[0:2] # Buffer is of length 2
       return jax.tree.map(_replace_with_zero_array, pytree)
     bsw = replace_with_zeros(self.layers.variables)
@@ -384,8 +384,8 @@ class Pipeline(nn.Module):
   def grab_bsw(self, vars, idx):
     def grab_bsw_leaf(leaf):
       arr = jax.lax.dynamic_slice_in_dim(leaf, idx, 1)
-      #return arr.astype(jnp.bfloat16) # TODO, should we modify where we change dtype? Unsure where
-      return arr
+      return arr.astype(jnp.bfloat16) # TODO, should we modify where we change dtype? Unsure where
+      #return arr
     return jax.tree.map(grab_bsw_leaf, vars)
 
   def force_ag(self, vars, sharding_info):
